ai_search/python/03_Upload_PDF_Permissions/requirements.txt
# Azure SDK packages
azure-search-documents>=11.4.0
azure-identity>=1.15.0
azure-core>=1.29.0

# Microsoft Graph SDK for Entra ID integration
msgraph-core>=1.0.0
msal>=1.24.0
requests>=2.31.0

# HTTP client for direct API calls
aiohttp>=3.9.0

# PDF processing
PyPDF2>=3.0.1
pdfplumber>=0.9.0

# Text processing and chunking
langchain>=0.1.0
langchain-text-splitters>=0.0.1
tiktoken>=0.5.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Logging and monitoring
structlog>=23.0.0
rich>=13.0.0

# Async support
asyncio-throttle>=1.0.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0
